Separate code into files
